Add jar explorer to file tree. Closes #4

+++ src/main/java/com/github/minecraft_ta/totalDebugCompanion/ui/components/treeView/FileTreeView.java
@@ -3,6 +3,7 @@ package com.github.minecraft_ta.totalDebugCompanion.ui.components.treeView;
 import com.formdev.flatlaf.extras.FlatSVGIcon;
 import com.github.minecraft_ta.totalDebugCompanion.CompanionApp;
 import com.github.minecraft_ta.totalDebugCompanion.lsp.JavaLanguageServer;
+import com.github.minecraft_ta.totalDebugCompanion.messages.codeView.DecompileAndOpenRequestMessage;
 import com.github.minecraft_ta.totalDebugCompanion.model.BaseScriptView;
 import com.github.minecraft_ta.totalDebugCompanion.model.CodeView;
 import com.github.minecraft_ta.totalDebugCompanion.model.ScriptView;
@@ -11,9 +12,14 @@ import com.github.minecraft_ta.totalDebugCompanion.ui.components.treeView.lazyFi
 import com.github.minecraft_ta.totalDebugCompanion.util.TextUtils;
 
 import javax.swing.*;
+import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.List;
 
 public class FileTreeView extends JScrollPane {
+
+    private static final Path MODS_PATH = CompanionApp.getRootPath().getParent().getParent().resolve("mods");
 
     public FileTreeView(EditorTabs tabs) {
         super();
@@ -21,6 +27,9 @@ public class FileTreeView extends JScrollPane {
         var tree = new LazyFileJTree() {
             @Override
             protected void showPopupMenu(LazyTreeNode node, TreeItem treeItem, int x, int y) {
+                if (!(treeItem instanceof FileSystemFileItem))
+                    return;
+
                 var popupMenu = new JPopupMenu();
                 var deleteItem = popupMenu.add("Delete");
                 deleteItem.setIcon(new FlatSVGIcon("icons/remove.svg"));
@@ -31,17 +40,29 @@ public class FileTreeView extends JScrollPane {
         };
 
         tree.addMouseDoubleClickListener((node, item) -> {
-            if (item.isDirectory() || !(item instanceof FileSystemFileItem fileItem))
+            if (item.isDirectory())
                 return;
 
-            if (node.getParent().getUserObject().getName().equals("src")) {
-                var name = fileItem.getName().replace(".java", "");
-                if (name.equals("BaseScript"))
-                    tabs.openEditorTab(new BaseScriptView(name));
-                else
-                    tabs.openEditorTab(new ScriptView(name));
-            } else {
-                tabs.openEditorTab(new CodeView(fileItem.getPath(), 1));
+            if (item instanceof FileSystemFileItem fileItem) {
+                if (node.getParent().getUserObject().getName().equals("src")) {
+                    var name = fileItem.getName().replace(".java", "");
+                    if (name.equals("BaseScript"))
+                        tabs.openEditorTab(new BaseScriptView(name));
+                    else
+                        tabs.openEditorTab(new ScriptView(name));
+                } else {
+                    tabs.openEditorTab(new CodeView(fileItem.getPath(), 1));
+                }
+            } else if (item instanceof ZipFileRootItem.Entry) {
+                if (!item.getName().endsWith(".class"))
+                    return;
+
+                //Reconstruct the class name
+                StringBuilder fullName = new StringBuilder(item.getName().substring(0, item.getName().length() - 6));
+                while (!((node = node.getParent()).getUserObject() instanceof ZipFileRootItem)) {
+                    fullName.insert(0, '.').insert(0, node.getUserObject().getName());
+                }
+                CompanionApp.SERVER.getMessageProcessor().enqueueMessage(new DecompileAndOpenRequestMessage(fullName.toString()));
             }
         });
 
@@ -60,7 +81,21 @@ public class FileTreeView extends JScrollPane {
 
         tree.addRootNodes(
                 tree.getItemFactory().createFileSystemDirectoryItem(JavaLanguageServer.SRC_DIR, true),
-                tree.getItemFactory().createFileSystemDirectoryItem(CompanionApp.getRootPath().resolve("decompiled-files"), true)
+                tree.getItemFactory().createFileSystemDirectoryItem(CompanionApp.getRootPath().resolve("decompiled-files"), true),
+                new DirectoryTreeItem("mods") {
+                    @Override
+                    public List<TreeItem> loadChildren() {
+                        try {
+                            return Files.walk(MODS_PATH, 1)
+                                    .skip(1)
+                                    .filter(p -> p.getFileName().toString().endsWith(".jar"))
+                                    .<TreeItem>map(ZipFileRootItem::new)
+                                    .toList();
+                        } catch (IOException e) {
+                            throw new RuntimeException(e);
+                        }
+                    }
+                }
         );
 
         setViewportView(tree);
